Added GPU gemm

+++ src/STC/Point.hs
@@ -1,8 +1,9 @@
 module STC.Point where
 
 data Point =
-  Point {-# UNPACK #-}!Int    -- x
-        {-# UNPACK #-}!Int    -- y
+  Point {-# UNPACK #-}!Double    -- x
+        {-# UNPACK #-}!Double    -- y
         {-# UNPACK #-}!Double --theta
         {-# UNPACK #-}!Double --scale
   deriving (Show, Read)
+
